requirements.txt
python-telegram-bot[job-queue]==21.4
aiohttp==3.9.5
beautifulsoup4==4.12.3
